(refactor): Reducing unnecessary precision of Point.color_rgb
(refactor): Renaming View methods

(fix): View::rotation formula and indicate the scalar order

* It does not affect the test results

+++ src/scene/point/mod.rs
@@ -5,6 +5,6 @@ pub use points::*;
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub struct Point {
     /// Normalized
-    pub color_rgb: [f64; 3],
+    pub color_rgb: [f32; 3],
     pub position: [f64; 3],
 }
